cleanup spacing and move UmfRouteEntry to its own file
